+++ src/auth/guards/index.ts
@@ -1,0 +1,2 @@
+export * from './jwt-auth.guard';
+export * from './local-auth.guard';
