+++ src/auth/strategies/index.ts
@@ -1,0 +1,2 @@
+export * from './jwt.strategy';
+export * from './local.strategy';
